Shooting weapon separated from entity
