+++ config/Config.py
@@ -2,7 +2,7 @@ import os
 
 
 WINDOW_RESOLUTION = (1280, 720)
-BACKGROUND_COLOR = (255, 255, 255)
+BACKGROUND_COLOR = (20, 20, 20)
 FPS = 60
 
 ENTITY_SPEED_FADE = 0.8
@@ -98,3 +98,7 @@ SWORD_DISTANCE = 30
 SWORD_COOLDOWN = 0.5
 SWORD_ALPHA = 90
 SWORD_DAMAGE = 1
+
+SHOOTING_WEAPON_SPRITE_PATH = os.path.join("pics", "white_rect.png")
+SHOOTING_WEAPON_DISTANCE = 30
+
