+++ objects/enemy/Turret.py
@@ -1,9 +1,9 @@
 from objects.enemy.Enemy import Enemy
 from config.Config import *
-from objects.weapon.Weapon import Weapon
+from objects.weapon.ShootingWeapon import ShootingWeapon
 
 
 class Turret(Enemy):
     def __init__(self, level, position):
         super().__init__(level, (level.visible, level.obstacle, level.entity), TURRET_PATH, position, 0, 0, TURRET_HEALTH, TURRET_ATTACK_RADIUS, TURRET_NOTICE_RADIUS)
-        self.equip_weapon(Weapon(level, self, 1.1 * BULLET_SPEED, 4 * BULLET_DAMAGE, BULLET_RANGE, BULLET_SPRITE_PATH, 4 * WEAPON_COOLDOWN))
+        self.equip_weapon(ShootingWeapon(level, SHOOTING_WEAPON_SPRITE_PATH, self, SHOOTING_WEAPON_DISTANCE, 4 * WEAPON_COOLDOWN, 1.1 * BULLET_SPEED, 4 * BULLET_DAMAGE, BULLET_RANGE, BULLET_SPRITE_PATH))
