+++ objects/friendly/Player.py
@@ -6,14 +6,14 @@ import pygame.math
 from config.Config import *
 from objects.main.Entity import Entity
 from objects.weapon.Bullet import Bullet
-from objects.weapon.Weapon import Weapon
+from objects.weapon.ShootingWeapon import ShootingWeapon
 from objects.weapon.Shield import Shield
 
 
 class Player(Entity):
     def __init__(self, level, groups, position):
         super().__init__(level, groups, PLAYER_SPRITE_PATH, position, PLAYER_ABS_ACCEL, PLAYER_MAX_SPEED, PLAYER_HEALTH, PLAYER_ENERGY) # move constants from config to __init__ (to create player with certain health, weapon, etc, in new location)
-        self.weapons = [Weapon(self.level, self, BULLET_SPEED, BULLET_DAMAGE, BULLET_RANGE, BULLET_SPRITE_PATH, WEAPON_COOLDOWN)]  # TODO: move to inventory later
+        self.weapons = [ShootingWeapon(self.level, SHOOTING_WEAPON_SPRITE_PATH, self, SHOOTING_WEAPON_DISTANCE, WEAPON_COOLDOWN, BULLET_SPEED, BULLET_DAMAGE, BULLET_RANGE, BULLET_SPRITE_PATH )]  # TODO: move to inventory later
         self.curr_weapon = 0    # index of self.weapons array
         self.shield = Shield(self.level, (self.level.shield, self.level.visible), SHIELD_SPRITE_PATH, self, SHIELD_DISTANCE, SHIELD_COOLDOWN)
