+++ objects/main/Solid.py
@@ -1,5 +1,5 @@
 import pygame
-from objects.weapon.Weapon import Weapon
+from objects.weapon.ShootingWeapon import Weapon
 from config.Config import *
 
 class Solid(pygame.sprite.Sprite):
