+++ objects/weapon/ColdSteel.py
@@ -1,7 +1,7 @@
-from objects.weapon.MeleeWeapon import MeleeWeapon
+from objects.weapon.Weapon import Weapon
 
 
-class ColdSteel(MeleeWeapon):
+class ColdSteel(Weapon):
     def __init__(self, level, groups, image_path, owner, owner_distance, cooldown, damage):
         super().__init__(level, groups, image_path, owner, owner_distance, cooldown)
         self.damage = damage
