+++ objects/weapon/Weapon.py
@@ -3,30 +3,41 @@ import pygame.math
 import pygame.sprite
 import pygame.rect
 from config.Config import *
-from objects.weapon.Bullet import Bullet
 
 
-class Weapon:
-    def __init__(self, level, owner, bullet_speed, bullet_damage, bullet_range, bullet_img_path, cooldown): # later <bullet_speed, ..., bullet_img_path> change to prepared Bullet examplar or to fabric
+class Weapon(pygame.sprite.Sprite):
+    def __init__(self, level, groups, image_path, owner, owner_distance, cooldown):
+        super().__init__(groups)
         self.level = level
+        self.start_image = pygame.image.load(image_path).convert_alpha()
+        self.image = pygame.image.load(image_path).convert_alpha()
+        self.rect = self.image.get_rect()
         self.owner = owner
-        self.bullet_speed = bullet_speed
-        self.bullet_damage = bullet_damage
-        self.bullet_range = bullet_range
-        self.bullet_img_path = bullet_img_path
-        self.last_shoot = cooldown
+        self.owner_distance = owner_distance
+        self.last_use = cooldown
         self.cooldown = cooldown
 
-    def update(self, dt):
-        self.last_shoot += dt
+        # check order of Level.bullets_update and Level.visible.update - self.uses is necessary because of this order
+        self.uses = [False, False]
 
-    def spawn_bullet(self, angle: pygame.math.Vector2):
-        if self.last_shoot >= self.cooldown:
-            self.last_shoot = 0
-            if angle.length() != 0:
-                angle = angle.normalize()
-            if hasattr(self.owner, "speed"):
-                speed = angle * (self.bullet_speed + angle.dot(self.owner.speed))   # added speed from owner, but in direction of shoot
-            else:
-                speed = angle * self.bullet_speed
-            self.level.bullets.add(Bullet(self.level, (self.level.bullets,), self.bullet_img_path, self.owner.pos, speed, self.bullet_damage, self.bullet_range, self))
+    def update(self, dt):
+        self.last_use += dt
+        self.move()
+        self.update_uses()
+
+    def move(self):
+        self.image = pygame.transform.rotate(self.start_image, self.owner.look_angle.angle_to(pygame.math.Vector2(0, 1)))
+        if not any(self.uses):
+            self.image.set_alpha(SHIELD_ALPHA)  # later add animation
+        self.rect = self.image.get_rect()
+        self.pos = self.owner.pos + self.owner.look_angle * self.owner_distance
+        self.rect.center = self.pos
+
+    def update_uses(self):
+        if all(self.uses):
+            self.uses = [False for _ in self.uses]
+        elif any(self.uses):
+            false_index = self.uses.index(False)
+            self.uses[false_index] = True
+
+
